Added the required files
index.py - allows users to select currency and duration to get the demographics with the risk indicator, min- max values, fxrate.py - Provides the FX rates for all currencies considering selected base currency, cb.py - It is custom currency basket to create and get the weight for user define currencies. NOTE : If we get the same webpage for the other funcaitonalities too, then kindly refresh the IDE, because all webpages are running on the same port number.